requirements.txt
google-genai
streamlit
youtube-transcript-api
python-dotenv